Completed Authentication

+++ routes/collab.js
@@ -5,7 +5,7 @@ const catchAsync = require("../utils/catchAsync");
 const ExpressError = require("../utils/ExpressError");
 const Collab = require("../models/unicollab"); //importing the DB Schema
 const collabSchema = require("../schemas");
-const { isLoggedIn } = require("../middleware");
+const { isLoggedIn, isAuthor } = require("../middleware");
 
 const validateCollab = (req, res, next) => {
   const { err } = collabSchema.validate(req.body);
@@ -43,10 +43,26 @@ router.get("/collab/new", isLoggedIn, (req, res) => {
 });
 
 router.get(
+  "/collab/mycollabs",
+  isLoggedIn,
+  catchAsync(async (req, res) => {
+    const user = req.user;
+    console.log(user);
+    const collabs = await Collab.find({ author: req.user._id });
+    console.log(collabs);
+    res.render("collab", { collabs });
+  })
+);
+
+router.get(
   "/collab/:id",
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const collab = await Collab.findById(id);
+    const collab = await Collab.findById(id).populate("author");
+    if (!collab) {
+      req.flash("error", "Cannot find the Collab!");
+      return res.redirect("/collab");
+    }
     res.render("show", { collab });
   })
 );
@@ -57,7 +73,9 @@ router.post(
   validateCollab,
   catchAsync(async (req, res, next) => {
     const collab = new Collab(req.body.collab);
+    collab.author = req.user._id;
     await collab.save();
+    req.flash("success", "Successfully Created a new Collab");
     res.redirect(`/collab/${collab._id}`);
   })
 );
@@ -65,9 +83,14 @@ router.post(
 router.get(
   "/collab/:id/edit",
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const { id } = req.params;
     const collab = await Collab.findById(id);
+    if (!collab) {
+      req.flash("error", "Cannot find the campground");
+      return res.redirect("/collab");
+    }
     res.render("edit", { collab });
   })
 );
@@ -75,12 +98,14 @@ router.get(
 router.put(
   "/collab/:id",
   isLoggedIn,
+  isAuthor,
   validateCollab,
   catchAsync(async (req, res) => {
     const { id } = req.params;
     const collab = await Collab.findByIdAndUpdate(id, {
       ...req.body.collab,
     });
+    req.flash("success", "Successfully Updated Collab");
     res.redirect(`/collab/${collab._id}`);
   })
 );
@@ -88,9 +113,16 @@ router.put(
 router.delete(
   "/collab/:id",
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const collab = await Collab.findById(id);
+    if (!collab.author.equals(req.user._id)) {
+      req.flash("error", "YOu do not have permission to do that!");
+      return res.redirect(`/collab/${id}`);
+    }
     await Collab.findByIdAndDelete(id);
+    req.flash("success", "Successfully deleted Collab");
     res.redirect("/collab");
   })
 );
